Emit non-BMP characters as \U escapes instead of malformed \x escapes (#2819)
The double-quoted scalar escape path assigned the character to a short
before testing it against 0xFF. char is unsigned 16-bit and short is
signed, so any character at or above U+8000 became negative and took the
'code <= 0xFF' branch. High surrogates are U+D800-U+DBFF, so the
IsHighSurrogate branch below it was unreachable.

The result was that (short)'\uD83D' is -10179 and
(-10179).ToString("X02") is "D83D", so the emitter wrote \xD83D. YAML's
\x escape takes exactly two hex digits, so re-parsing that read \xD8
followed by a literal "3D" and the character was silently corrupted:
U+1F600 round-tripped to 'Ø3DÞ00'.

Widening code to int makes the 0xFF test mean what it reads as and lets
the existing surrogate-pair branch emit \U0001F600.

+++ src/Meziantou.Framework.Yaml/Emitter.cs
@@ -1199,7 +1199,7 @@ public partial class Emitter : IEmitter
                         break;
 
                     default:
-                        short code = (short)character;
+                        int code = character;
                         if (code <= 0xFF)
                         {
                             Write('x');
